Refactored a fair bit, still more to do - primarily remains to wrap all functionality into classes
