+++ backend/api/transcription_job/types.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class TranscriptionJob(BaseModel):
+    podcast_id: str
+    episode_number: str
