+++ backend/main.py
@@ -1,11 +1,13 @@
 from modal import Mount, asgi_app
 
-from backend.app.functions import stub, APP_VOLUME
+from backend.ops.stub import stub
+from backend.ops.storage import APP_VOLUME
 from backend import config
 
 # think you must import all modal operator functions here (might want to create a class that discovers them, as long as in files of same name in each api subdirectory)
 
-from backend.api.podcast.operators import * # could also be in podcast folder you use __init__.py to import all operators into there, and then here you just import the podcast package/folder (but long term best to make them auto-discoverable)
+from backend.ops.podcast.operators import * # could also be in podcast folder you use __init__.py to import all operators into there, and then here you just import the podcast package/folder (but long term best to make them auto-discoverable)
+from backend.ops.transcription_job.operators import *
 
 
 @stub.function(
@@ -17,11 +19,9 @@ from backend.api.podcast.operators import * # could also be in podcast folder yo
 @asgi_app()
 def fastapi_app():
     import fastapi.staticfiles
-
-    from backend.api.main import web_app # NOTE: this is what makes all the routes in the api.py visible/available to the modal frontend
-
+    from backend.api.main import web_app 
+    # NOTE: this import is what makes all the routes in `api` available to the modal frontend
     web_app.mount(
         "/", fastapi.staticfiles.StaticFiles(directory="/assets", html=True)
     )
-
     return web_app
