+++ backend/ops/stub.py
@@ -1,0 +1,10 @@
+import modal
+
+from backend.ops.image import APP_IMAGE
+
+stub = modal.Stub(
+    "whisper-pod-transcriber",
+    image=APP_IMAGE
+)
+
+stub.in_progress = modal.Dict.new()
