+++ backend/ops/transcription_job/constants.py
@@ -1,0 +1,2 @@
+
+MAX_JOB_AGE_SECS = 10 * 60
