+++ backend/ops/types.py
@@ -1,0 +1,5 @@
+from typing import NamedTuple
+
+class InProgressJob(NamedTuple):
+    job_id: str
+    start_time: int
