+++ backend/src/podcast/constants.py
@@ -1,0 +1,10 @@
+from .types import PodcastMetadata
+
+THE_PODCAST = PodcastMetadata(
+    id="twiml-ai-podcast",
+    title="The TWIML AI Podcast",
+    description="description",
+    html_description="htmlDescription",
+    web_url="https://twimlai.com",
+    feed_url="https://twimlai.com/feed/"
+)
