+++ backend/src/podcast/functions/download.py
@@ -1,0 +1,46 @@
+import urllib.request
+import pathlib
+import urllib.request
+
+from backend import config
+from .helpers import sizeof_fmt
+
+logger = config.get_logger(__name__)
+
+from ..types import DownloadResult
+
+def download_podcast_file(url: str) -> DownloadResult:
+    req = urllib.request.Request(
+        url,
+        data=None,
+        # Set a user agent to avoid 403 response from some podcast audio servers.
+        headers={
+            "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9_3) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/35.0.1916.47 Safari/537.36"
+        },
+    )
+    with urllib.request.urlopen(req) as response:
+        return DownloadResult(
+            data=response.read(),
+            content_type=response.headers["content-type"],
+        )
+    
+def store_original_audio(
+    url: str, destination: pathlib.Path, overwrite: bool = False
+) -> None:
+    if destination.exists():
+        if overwrite:
+            logger.info(
+                f"Audio file exists at {destination} but overwrite option is specified."
+            )
+        else:
+            logger.info(
+                f"Audio file exists at {destination}, skipping download."
+            )
+            return
+
+    podcast_download_result = download_podcast_file(url=url)
+    humanized_bytes_str = sizeof_fmt(num=len(podcast_download_result.data))
+    logger.info(f"Downloaded {humanized_bytes_str} episode from URL.")
+    with open(destination, "wb") as f:
+        f.write(podcast_download_result.data)
+    logger.info(f"Stored audio episode at {destination}.")
