+++ backend/src/podcast/functions/helpers.py
@@ -1,3 +1,10 @@
+def sizeof_fmt(num, suffix="B") -> str:
+    for unit in ["", "Ki", "Mi", "Gi", "Ti", "Pi", "Ei", "Zi"]:
+        if abs(num) < 1024.0:
+            return "%3.1f%s%s" % (num, unit, suffix)
+        num /= 1024.0
+    return "%.1f%s%s" % (num, "Yi", suffix)
+
 def sorting_key(item):
     episode_number = item[1].get("episode_number")
     if episode_number is None:
