+++ backend/src/podcast/functions/segments.py
@@ -1,0 +1,31 @@
+from ..types import Segment
+
+def coalesce_short_transcript_segments(
+    segments: list[Segment],
+) -> list[Segment]:
+    """
+    Some extracted transcript segments from openai/whisper are really short, like even just one word.
+    This function accepts a minimum segment length and combines short segments until the minimum is reached.
+    """
+    minimum_transcript_len = 200  # About 2 sentences.
+    previous = None
+    long_enough_segments = []
+    for current in segments:
+        if previous is None:
+            previous = current
+        elif len(previous["text"]) < minimum_transcript_len:
+            previous = _merge_segments(left=previous, right=current)
+        else:
+            long_enough_segments.append(previous)
+            previous = current
+    if previous:
+        long_enough_segments.append(previous)
+    return long_enough_segments
+
+
+def _merge_segments(left: Segment, right: Segment) -> Segment:
+    return {
+        "text": left["text"] + " " + right["text"],
+        "start": left["start"],
+        "end": right["end"],
+    }
