+++ backend/src/transcription_job/functions/helpers.py
@@ -1,0 +1,4 @@
+import datetime
+
+def utc_now() -> datetime.datetime:
+    return datetime.datetime.now(datetime.timezone.utc)
