+++ backend/src/transcription_job/functions/silences.py
@@ -1,0 +1,56 @@
+from typing import Iterator, Tuple
+
+from backend import config
+
+logger = config.get_logger(__name__)
+
+def split_silences(
+    path: str, min_segment_length: float = 30.0, min_silence_length: float = 1.0
+) -> Iterator[Tuple[float, float]]:
+    """Split audio file into contiguous chunks using the ffmpeg `silencedetect` filter.
+    Yields tuples (start, end) of each chunk in seconds."""
+
+    import re
+
+    import ffmpeg
+
+    silence_end_re = re.compile(
+        r" silence_end: (?P<end>[0-9]+(\.?[0-9]*)) \| silence_duration: (?P<dur>[0-9]+(\.?[0-9]*))"
+    )
+
+    metadata = ffmpeg.probe(path)
+    duration = float(metadata["format"]["duration"])
+
+    reader = (
+        ffmpeg.input(str(path))
+        .filter("silencedetect", n="-10dB", d=min_silence_length)
+        .output("pipe:", format="null")
+        .run_async(pipe_stderr=True)
+    )
+
+    cur_start = 0.0
+    num_segments = 0
+
+    while True:
+        line = reader.stderr.readline().decode("utf-8")
+        if not line:
+            break
+        match = silence_end_re.search(line)
+        if match:
+            silence_end, silence_dur = match.group("end"), match.group("dur")
+            split_at = float(silence_end) - (float(silence_dur) / 2)
+
+            if (split_at - cur_start) < min_segment_length:
+                continue
+
+            yield cur_start, split_at
+            cur_start = split_at
+            num_segments += 1
+
+    # silencedetect can place the silence end *after* the end of the full audio segment.
+    # Such segments definitions are negative length and invalid.
+    if duration > cur_start and (duration - cur_start) > min_segment_length:
+        yield cur_start, duration
+        num_segments += 1
+    logger.info(f"Split {path} into {num_segments} segments")
+
